fix: Specifications from parents of Sphinx source directory (#264)
* fix: Specifications from parents of source dir

If a spec is referenced with `../../../and-so-on/`
to access a file from a parent of the Sphinx
source directory, the resolved path can be
a path that is a parent of the output HTML directory.

By replacing the ".." with "dot-dot" in the output
of the Sphinx relfn2path function, this fix
ensures that the spec is copied correctly to
a subdirectory of the output `_static` directory.

For example, expect to see `_static/dot-dot/dot-dot/openapi.yaml`.
That path would otherwise be a parent of the output HTML.

Signed-off-by: Mike McKiernan <mmckiernan@nvidia.com>

* chore: Update version for patch release

Signed-off-by: Mike McKiernan <mmckiernan@nvidia.com>

---------

Signed-off-by: Mike McKiernan <mmckiernan@nvidia.com>